AG-9098 - Fix circular dependency.
